move GlobalExceptionHandler to exception in customer-service
